Refactor: extract command handlers from main.rs into commands/ modules
Break up the 5,729-line main.rs monolith by extracting all 48 command
handlers into focused modules under src/commands/:

- commands/control.rs (1,112 lines): on, off, blind, light, gate,
  thermostat, alarm, door, intercom, charger, music, lock/unlock, etc.
- commands/inspect.rs (865 lines): ls, get, info, rooms, categories,
  globals, modes, sensors, energy, weather, stats, history, watch,
  stream, if, autopilot
- commands/system.rs (853 lines): status, log, time, discover,
  extensions, health, update, reboot, files, otel
- commands/config_cmd.rs (762 lines): setup, alias, scene, cache,
  token, config, completions, schema

main.rs is now 2,398 lines (58% reduction) — a thin dispatcher with
CLI definitions, helper functions, and the match-based routing table.

All 260 tests pass. Zero clippy warnings.
